deprecate new_window and new_float commands
